KS-2-2 update task_2

+++ src/main/kotlin/lesson2/lesson2_task1.kt
@@ -5,11 +5,12 @@ fun main() {
     val englishGradeofStudent2 = 4
     val englishGradeofStudent3 = 3
     val englishGradeofStudent4 = 5
-    val numberOfStudents = (4).toFloat()
+
+    val numberOfStudents = 4f
 
     val averageScoreInEnglish = (
-            englishGradeofStudent1 + englishGradeofStudent2 +
-            englishGradeofStudent3 + englishGradeofStudent4) / numberOfStudents
+        englishGradeofStudent1 + englishGradeofStudent2 +
+        englishGradeofStudent3 + englishGradeofStudent4) / numberOfStudents
 
     println(averageScoreInEnglish)
 }
